Implement signed strides for zero-copy flip operations
Layout now uses Vec<isize> for strides, enabling negative strides for
zero-copy flip. The flip operation becomes O(1) metadata-only instead
of O(n) element copying.

Changes:
- Layout: strides now Vec<isize>, added flip() method
- StridedIter: uses isize for storage_index arithmetic
- EmberTensor: added with_layout() for zero-copy view operations
- flip.rs: simplified to 3-line zero-copy implementation
- Fixed stride arithmetic in binary, comparison, reduce, slice ops
- Documented signed strides in ARCHITECTURE.md

+++ crates/burn-ember/src/ops/binary.rs
@@ -199,17 +199,17 @@ where
 {
     let (rows, cols, l_row_stride, l_col_stride) = lhs_layout.as_2d_strides().unwrap();
     let (_, _, r_row_stride, r_col_stride) = rhs_layout.as_2d_strides().unwrap();
-    let l_offset = lhs_layout.start_offset();
-    let r_offset = rhs_layout.start_offset();
+    let l_offset = lhs_layout.start_offset() as isize;
+    let r_offset = rhs_layout.start_offset() as isize;
 
     let mut result = Vec::with_capacity(rows * cols);
 
     for row in 0..rows {
-        let l_row_start = l_offset + row * l_row_stride;
-        let r_row_start = r_offset + row * r_row_stride;
+        let l_row_start = l_offset + row as isize * l_row_stride;
+        let r_row_start = r_offset + row as isize * r_row_stride;
         for col in 0..cols {
-            let l_idx = l_row_start + col * l_col_stride;
-            let r_idx = r_row_start + col * r_col_stride;
+            let l_idx = (l_row_start + col as isize * l_col_stride) as usize;
+            let r_idx = (r_row_start + col as isize * r_col_stride) as usize;
             result.push(op(lhs[l_idx], rhs[r_idx]));
         }
     }
